fixed `assemble`
